Make SDK search docs more clear (#228)
After chatting with @richardgill and a user on discord over the last
day, I realized our search docs didn't do a great job explaining certain
core features. Specifically the highlight examples weren't very clear.

Using Keanu Reeves is a good example for the typo tolerance examples,
but using `Users` and `Posts` along with that term just confused the
examples (especially when you're filtering posts by a `label` with a
value of `movies` :dizzy_face:). I rewrote the tables to be `Movies` and
`Actors`.

I also went and added line highlights to all the examples and fixed a
lot of the code (which were referencing object columns :facepalm: ).

---------

Co-authored-by: Philip Krauss <philip@philipkrauss.at>