Deemphasizes workflow in prominence in the docs (#251)
This is a draft PR for discussion with @richardgill / @kostasb /
@alexfrancoeur

Our workflow has limitations that make difficult to lead with as a
feature offering. Within the docs, this PR:

- Removes workflow's prominence in our docs, and adds an alert to the
top of the workflow documentation
- Moves the available regions page outside of getting started, and moves
it to resources (for now).

No slugs are changed. URLs will continue to work.

Is paired with https://github.com/xataio/frontend-next/pull/2921 on the
website