Added deployment explanation to docs for the framework starter guide for Astro (#329)
Added a deployment explanation for the framework starter guide for
Astro. That explains the adding of a ssr adapter for deployment use. Of
a Astro Xata blog, from the starter guide. Also fixed a typo with the
naming of a component from src/app/page.tsx to src/pages/index.astro. To
help with confusion of naming components or where said code should go
while adding the search functionality to the app.