add ssl require option to postgres docs (#336)
Adds sslmode require option in all examples, even if it's redundant for
psql which defaults to it.

The connection string requires this option to work with other clients
and ORMs (i.e. Drizzle) and it's hard for a user to notice the
difference, so it's best to promote syntax that covers more uses.

From user feedback on
[Discord](https://discord.com/channels/996791218879086662/996791219348852836/1245718350479233046).