Fixes for summarize and aggregate docs (#350)
- Fix incorrect call syntax for columns in summarize snippets
- Rephrase suggestion to using Aggregate as a "no result limits"
endpoint as it is not true for every call. For example, a common case
for users trying out aggregate is the 1000 result buckets of Summarize,
which still applies in Aggregate's [top
values](https://xata.io/docs/sdk/aggregate#top-values). This avoids
setting expectations that can't be met in every case.